Create import_poi_data_lightning.py
An optimised version of the _fast algorithm, removed unnecessary loops and only using dual threading. Also de-coupled loops from file reading in json and csv and significantly reduced processing time of data 50% gains on fast algorithm.
This is final as I do not think it is possible to optimise the algorithm any further in python to be any faster, next step would be to use a C++ or C binding to do the heavy lifting and a different DB engine to allow concurrency.